third_party/WebKit/LayoutTests/external/wpt/html/infrastructure/urls/resolving-urls/query-encoding/location.sub_encoding=x-cp1251-expected.txt

This is a testharness.js-based test.
FAIL location [PutForwards] assert_equals: expected "?&%23255;" but got "?%26%23255%3B"
FAIL location.assign() assert_equals: expected "?&%23255;" but got "?%26%23255%3B"
FAIL location.replace() assert_equals: expected "?&%23255;" but got "?%26%23255%3B"
FAIL location.href assert_equals: expected "?&%23255;" but got "?%26%23255%3B"
FAIL location.search assert_equals: expected "?&%23255;" but got "?%C3%BF"
Harness: the test ran to completion.
